move some test classes to test fixtures
