+++ common4j/src/testFixtures/java/com/microsoft/identity/http/ResponseBody.java
@@ -20,7 +20,7 @@
 //  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //  THE SOFTWARE.
-package com.microsoft.identity.common.java.net.util;
+package com.microsoft.identity.http;
 
 public class ResponseBody {
